Creados -DTO para obtener todos los registros de entradas-salidas de la BD -Repositorio para tabla registro -Repositorio para la tabla tarjeta -Service para registros

+++ API/acces-control/src/main/java/com/controlacceso/accescontrol/DTO/RegistroEntradasSalidasDTO.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public record RegistroEntradasSalidasDTO(
         String empleado,
-        Long idEmpleado,
+        Integer idEmpleado,
         String numeroTarjeta,
         List<RegistrosEntradasSalidasDetalleDTO> registros
 ) {
