+++ API/acces-control/src/main/java/com/controlacceso/accescontrol/repository/TarjetaRepository.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 @Repository
 public interface TarjetaRepository extends JpaRepository<Tarjeta, Integer> {
     Optional<Tarjeta> findByUid(String uid);
-    Optional<Tarjeta> findByEmpleadoId(Long empleadoId);
+    Optional<Tarjeta> findByEmpleadoId(Integer empleadoId);
 }
